fix: main.tsxをルートにするように修正
