+++ index.html
@@ -12,6 +12,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/app.tsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
