+++ src/app.tsx
@@ -1,11 +1,6 @@
 import { RouterProvider, createRouter } from "@tanstack/react-router";
-import { StrictMode } from "react";
-import ReactDOM from "react-dom/client";
 import "./index.css";
 import { routeTree } from "./routeTree.gen";
-import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "./pkg/api/client/queryClient";
-
 const router = createRouter({ routeTree });
 
 declare module "@tanstack/react-router" {
@@ -14,14 +9,8 @@ declare module "@tanstack/react-router" {
   }
 }
 
-const rootElement = document.getElementById("app")!;
-if (!rootElement.innerHTML) {
-  const root = ReactDOM.createRoot(rootElement);
-  root.render(
-    <StrictMode>
-      <QueryClientProvider client={queryClient}>
-        <RouterProvider router={router} />
-      </QueryClientProvider>
-    </StrictMode>,
-  );
+function App() {
+  return <RouterProvider router={router} />;
 }
+
+export default App;
