+++ src/main.tsx
@@ -1,0 +1,18 @@
+import { StrictMode } from "react";
+import ReactDOM from "react-dom/client";
+import "./index.css";
+import { QueryClientProvider } from "@tanstack/react-query";
+import { queryClient } from "./pkg/api/client/queryClient";
+import App from "./app";
+
+const rootElement = document.getElementById("app")!;
+if (!rootElement.innerHTML) {
+  const root = ReactDOM.createRoot(rootElement);
+  root.render(
+    <StrictMode>
+      <QueryClientProvider client={queryClient}>
+        <App />
+      </QueryClientProvider>
+    </StrictMode>,
+  );
+}
